# optimized make_sequence with gcc