```cpp
using list = brigand::make_sequence<std::integral_constant<int,0>, SIZE>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 35M | 00.08s - 36M | 00.09s - 37M | 00.12s - 39M | 00.18s - 43M
gcc | 00.05s - 20M | 00.07s - 23M | 00.10s - 27M | 00.15s - 35M | 00.27s - 55M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 35M | 00.07s - 36M | 00.09s - 37M | 00.12s - 39M | 00.18s - 43M
gcc | 00.05s - 20M | 00.06s - 22M | 00.09s - 26M | 00.13s - 32M | 00.22s - 46M